Report (pdf file)
Capstone Project Overview - The Battle of Neighborhoods (Week 2)

Existing Problem
Our client, Coffee Express Ltd., is an SME specializing on serving coffee and beverages including Latte, Cappuccino, Espresso and hot chocolate at Downtown Vancouver in Canada. Recently, they approach to us and express their interests in expending their branches in a city of the world’s leading financial center in North America, in particular to either New York or Toronto. Both cities are the financial capitals with many skyscrapers and businessmen where the local demands of morning beverages remain high. However, it is difficult for Coffee Express Ltd. to make the decision on the best location for their expansion since they are not familiar with the demographics in both cities. 

Data Description
We can gather the neighborhoods data from Foursquare API and apply K-means clustering to prioritize the best coffee shop locations in either one of the cities based on the foot traffic. Finally, we can conclude the best city and location for Coffee Express Ltd. to start their expansion. 

Methodologies
From Foursquare API, we gather the distribution of the nearest 30 potential competitors that running similar business as our client to determine whether they pose a great threat to our client. The attached map visualization centers at the Toronto Financial District, where the commercial area and skyscrapers are, and shows that the distribution of the 30 cafés operating around there. 

While the Toronto Financial District is a symbolic representation of the financial district at Toronto, we further stimulate the World Trade Center at Manhattan, New York to be another possible area for our client to start their expansion there. We further visualize the distribution of the nearest 30 local competitors near the World Trade Center for our further analysis.
With the same scale map visualization between Toronto and New York, we can interpret that the competition in New York is more intense than in Toronto. For the competitor distribution in New York, the distances between each store and our center is much shorter than in Toronto. Assuming the same foot traffic between the Toronto Financial District and World Trade Center at Manhattan, we can expect that the cafe in New York is facing relatively lower average sales per customer and lower repeat purchase rate as customers have more choices nearby in New York. 

Conclusion
As we can see from the Foursquare API, the financial district in Toronto has a lower density of local competitor distribution than in New York, which indicates that our client do not have to face a vicious or cut-throat price competition with the local cafés and, thus, can focus on improving their product and service quality comparing to New York under the same resources. Therefore, we suggest Coffee Express Ltd. to start their expansion in the financial district at Toronto instead of New York.